Update the version of less
Update the version of less to support new modern features like `Properties as Variables` as described [here](http://lesscss.org/features/#variables-feature-properties-as-variables-new-). I went through the code and tested it that I just need to update the version defined in `package.json` to make it works. Any other code do not need to modify.